Move logos under assets/ in lib-customizations
- All image type files should be under assets unless they have a good reason to be under a more specific path
