+++ frontend/src/lib-customizations/espoo/employee.tsx
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
 import type { EmployeeCustomizations } from 'lib-customizations/types'
-import EspooLogo from './EspooLogo.png'
+import EspooLogo from './assets/EspooLogo.png'
 import featureFlags from './featureFlags'
 
 const customizations: EmployeeCustomizations = {
